Removing report_dataset() from flattener, instead utilizing get_metdata(). Moving get_value() to gather.py. Making comments look prettier.
